Removed Pandoc-breaking BibTeX entry
Zotero made a chance where one of the bibliography entries was exported with the name "office_of_research_&_development_predictive_nodate". The "&" character is illegal as a naming character, which broke the Pandoc build process.

I've replaced it with a better reference to the same paper, which fully specifies the author names.