Fixes #82
The underlying issue was LaTeX thought the best location for the table was at the end of the document. This would then place it in the appenix, also giving it the appendix table numbering.

`kable()` invoked `longtable` in the resulting LaTeX, which keeps the table in the same location of the code. However, the use of kableExtra seemed to invoke the simple `table` in LaTeX, which floats the table.

Interestingly, `kable_styling(latex_options = "hold_position")` did not work (which placed the [!h] option in LaTeX). However, invoking `longtable` in the `kable()` command did work.

I've also swapped `setwd()` with `opts_knit$set(root.dir = ...)` per the documentation. This was not the cause of the issue, but did it to eliminate the possibility.